backend/consumer/requirements.txt
kafka-python==2.0.2
opencv-python==4.5.4.58
numpy